Add files via upload
Updated a few components and styling to fix compatibility issues with the build version of the app. For some reason while compiling css, create-react-app will convert nn% opacity to 1%, making elements almost inivisible. A workaround for this is to use n.n as an alternaitve to percentiles (such as opacity: 0.4; opacity:0.9; opacity: 1;)